Let keyboard broadcast the keyevents to kmmanager.
Additionally, the architecture-independent stuff has been moved
into a library.

+++ source/drivers/eco32/keyboard/set2.h
@@ -29,4 +29,4 @@
  * @param scanCode the received scancode
  * @return true if it was a keycode
  */
-bool kb_set2_getKeycode(uchar *isBreak,uchar *keycode,uchar scanCode);
+bool kb_set2_getKeycode(uchar *isBreak,uchar *keycode,uint32_t scanCode);
